requirements.txt
Django==1.7.8
wsgiref==0.1.2
Pillow==2.9.0
djangorestframework==3.1.3